[Refactor] Moved reset_top_left_panel_bottom_row() to common_ui.c/.h
