+++ include/game/common_ui.h
@@ -33,4 +33,9 @@ enum BackgroundId get_current_background(void);
  */
 void change_background(enum BackgroundId id, bool force_redraw);
 
+/**
+ * @brief Restores the bottom row of the top-left panel from the background map.
+ */
+void reset_top_left_panel_bottom_row(void);
+
 #endif // COMMON_UI_H
